Rename and move interface code from `main.cpp`
